Create validate.rs
