+++ raphtory-graphql/tests/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod graphql;
+pub mod validate;
 pub mod jwt;
 pub mod strategy;
 pub mod tree;
